+++ crates/rusterix/src/vm/compile.rs
@@ -825,6 +825,25 @@ impl Visitor for CompileVisitor {
                             loc,
                         ));
                     }
+                } else if name == "dialog" {
+                    if args.len() == 2 {
+                        for arg in args {
+                            _ = arg.accept(self, ctx)?;
+                        }
+                        ctx.emit(NodeOp::HostCall {
+                            name: "dialog".into(),
+                            argc: args.len() as u8,
+                        });
+                    } else {
+                        return Err(RuntimeError::new(
+                            format!(
+                                "Wrong amount of arguments for '{}', expected '2' got '{}'",
+                                name,
+                                args.len(),
+                            ),
+                            loc,
+                        ));
+                    }
                 } else if name == "patrol" {
                     if args.len() <= 2 {
                         for arg in args {